# Requests/Legalbot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Legalbot API",
    "version": "1.0.0",
    "description": "OpenAPI export for the Legalbot backend endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:9090"
    }
  ],
  "paths": {
...
    "/insert": {
      "post": {
        "summary": "Create a suit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "List suits for a lawyer",
        "parameters": [
          {
            "name": "lawyer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of suits"
          }
        }
      }
    },
...
    "/api/documents/templates/{templateKey}": {
      "get": {
        "summary": "Get a document template",
        "parameters": [
          {
            "name": "templateKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template returned"
          }
        }
      }
    },
...
    "/api/documents/drafts/{draftId}": {
      "get": {
        "summary": "Get draft context",
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft retrieved"
          }
        }
      }
    }
  }
}